Updated script with min vessel size options
Added an option to choose a minimum vessel size, as this has been requested quite a few times on the workshop page. 
Aircraft have now been excluded from the size restrictions as they have their own option to be toggled on/off and only really come in size small anyway.