re-added the setNPCSeats command from Combat State into a different section
Added the setNPCSeats command that was commented out back into the script but in the same function that updates the IFF frequency, also increased the timer to 25 seconds instead of 6.

This was due to the fact that making the crew re-sit back into their seats every so often is extremely useful, as it also inadvertently fixes the bug where sometimes the Gunner AI turns the designator too far, causing it to glitch out. Popping in and out of the gunner's seat makes it reset to zero and behave as normal again.