Update script with markers changes
Added a completely separate option to show the victim markers, i.e. the ones that show up on the map whenever a civilian ship or aircraft gets close to one of the hostiles, as quite a few players were asking for the option to have them both off or on at once.